extract and cleanup
